requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dominate==2.9.1
einops==0.8.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.66.2
h5py==3.11.0
huggingface-hub==0.25.2
idna==3.7
imagecodecs==2024.9.22
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
monai==1.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nystrom-attention==0.0.12
opencv-python==4.10.0.84
openslide-python==1.3.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pycparser==2.22
pyparsing==3.1.2
pytorch-lightning==2.4.0
pytorch-toolbelt==0.6.3
pytz==2024.1
pyvips==2.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sympy==1.13.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==1.0.9
tokenizers==0.21.0
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.47.1
triton==3.0.0
tzdata==2024.1
urllib3==2.2.3
v==1
visdom==0.2.4
websocket-client==1.8.0
Werkzeug==3.0.4
yarl==1.9.4
